+++ test/test_wiki.rb
@@ -61,30 +61,6 @@ context "Wiki" do
       @wiki.normalize_commit(commit.dup))
   end
 
-  #test "#tree_map_for caches ref and tree" do
-  #  assert @wiki.ref_map.empty?
-  #  assert @wiki.tree_map.empty?
-  #  @wiki.tree_map_for 'master'
-  #  assert_equal({"master"=>"60f12f4254f58801b9ee7db7bca5fa8aeefaa56b"}, @wiki.ref_map)
-  #
-  #  map = @wiki.tree_map['60f12f4254f58801b9ee7db7bca5fa8aeefaa56b']
-  #  assert_equal 'Bilbo-Baggins.md',        map[0].path
-  #  assert_equal '',                        map[0].dir
-  #  assert_equal map[0].path,               map[0].name
-  #  assert_equal 'Mordor/Eye-Of-Sauron.md', map[3].path
-  #  assert_equal '/Mordor',                 map[3].dir
-  #  assert_equal 'Eye-Of-Sauron.md',        map[3].name
-  #end
-  #
-  #test "#tree_map_for only caches tree for commit" do
-  #  assert @wiki.tree_map.empty?
-  #  @wiki.tree_map_for '60f12f4254f58801b9ee7db7bca5fa8aeefaa56b'
-  #  assert @wiki.ref_map.empty?
-  #
-  #  entry = @wiki.tree_map['60f12f4254f58801b9ee7db7bca5fa8aeefaa56b'][0]
-  #  assert_equal 'Bilbo-Baggins.md', entry.path
-  #end
-
   test "text_data" do
     wiki = Gollum::Wiki.new(testpath("examples/yubiwa.git"))
     if String.instance_methods.include?(:encoding)
@@ -96,6 +72,19 @@ context "Wiki" do
       page = wiki.page("strider")
       assert_equal page.raw_data, page.text_data
     end
+  end
+
+  test "gets reverse diff" do
+    diff = @wiki.full_reverse_diff('a8ad3c09dd842a3517085bfadd37718856dee813')
+    assert_match "b/Mordor/_Sidebar.md", diff
+    assert_match "b/_Sidebar.md", diff
+  end
+
+  test "gets reverse diff for a page" do
+    diff  = @wiki.full_reverse_diff_for('_Sidebar.md', 'a8ad3c09dd842a3517085bfadd37718856dee813')
+    regex = /b\/Mordor\/\_Sidebar\.md/
+    assert_match    "b/_Sidebar.md", diff
+    assert_no_match regex, diff
   end
 end
 
@@ -290,6 +279,39 @@ context "Wiki sync with working directory" do
     page = @wiki.page("New Page")
     @wiki.delete_page(page, commit_details)
     assert !File.exist?(File.join(@path, "New-Page.md"))
+  end
+
+  teardown do
+    FileUtils.rm_r(@path)
+  end
+end
+
+context "page_file_dir option" do
+  setup do
+    @path = cloned_testpath('examples/page_file_dir')
+    @repo = Grit::Repo.init(@path)
+    @page_file_dir = 'docs'
+    @wiki = Gollum::Wiki.new(@path, :page_file_dir => @page_file_dir)
+  end
+
+  test "write a page in sub directory" do
+    @wiki.write_page("New Page", :markdown, "Hi", commit_details)
+    assert_equal "Hi", File.read(File.join(@path, @page_file_dir, "New-Page.md"))
+    assert !File.exist?(File.join(@path, "New-Page.md"))
+  end
+
+  test "a file in page file dir should be found" do
+    assert @wiki.page("foo")
+  end
+
+  test "a file out of page file dir should not be found" do
+    assert !@wiki.page("bar")
+  end
+
+  test "search results should be restricted in page filer dir" do
+    results = @wiki.search("foo")
+    assert_equal 1, results.size
+    assert_equal "foo", results[0][:name]
   end
 
   teardown do